pyproject.toml
[project]
name = "claude-proxy-bridge"
version = "1.0.0"
description = "Reverse proxy bridge exposing Claude Code CLI as OpenAI-compatible API endpoints"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
claude-proxy-bridge = "start:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
